requirements.txt
asgiref==3.7.2
boto3==1.28.21
botocore==1.31.21
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3